requirements.txt
alabaster==0.7.16
babel==2.16.0
blinker==1.8.2
blurb==1.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docutils==0.21.2
Flask==3.0.3
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
Pygments==2.18.0
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-doc==2020.10.1
python-docs-theme==2024.6
python-docx==1.1.2
pytz==2024.2
requests==2.32.3
requests-mock==1.12.1
setuptools==75.2.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==2.4.4
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
